Timing mossa corretto (massimizza il tempo a meno che non ottiene una mossa eccellente)

+++ MyAgente/src/it/jar/mulino/model/Mossa.java
@@ -29,15 +29,15 @@ public class Mossa {
         this(player, Byte.MAX_VALUE, to, Byte.MAX_VALUE);
     }
 
-    public int getTo() {
+    public byte getTo() {
         return to;
     }
 
-    public int getRemove() {
+    public byte getRemove() {
         return remove;
     }
 
-    public int getFrom() {
+    public byte getFrom() {
         return from;
     }
